+++ Legacy/classes.php
@@ -4,10 +4,14 @@
 interface Tx_Rnbase_Configuration_ProcessorInterface extends \Sys25\RnBase\Configuration\ConfigurationInterface
 {
 }
+echo __LINE__.'<br>';
+
 /** @deprecated */
 class tx_rnbase_util_TYPO3 extends Sys25\RnBase\Utility\TYPO3
 {
 }
+echo __LINE__.'<br>';
+
 /** @deprecated */
 class tx_rnbase_configurations extends Sys25\RnBase\Configuration\Processor
 {
@@ -60,10 +64,6 @@ class Tx_Rnbase_Service_Authentication extends \Sys25\RnBase\Typo3Wrapper\Servic
 }
 /** @deprecated */
 class Tx_Rnbase_Service_Base extends \Sys25\RnBase\Typo3Wrapper\Service\AbstractService
-{
-}
-/** @deprecated */
-abstract class Tx_Rnbase_Repository_AbstractRepository extends \Tx_Rnbase_Domain_Repository_AbstractRepository
 {
 }
 /** @deprecated */
@@ -135,3 +135,59 @@ class tx_rnbase_util_Misc extends Sys25\RnBase\Utility\Misc
 class tx_rnbase_util_Network extends Sys25\RnBase\Utility\Network
 {
 }
+/** @deprecated */
+class Tx_Rnbase_Domain_Collection_Base extends Sys25\RnBase\Domain\Collection\BaseCollection
+{
+}
+/** @deprecated */
+interface Tx_Rnbase_Domain_Model_DataInterface extends Sys25\RnBase\Domain\Model\DataInterface
+{
+}
+/** @deprecated */
+interface Tx_Rnbase_Domain_Model_DomainInterface extends Sys25\RnBase\Domain\Model\DomainInterface
+{
+}
+/** @deprecated */
+interface Tx_Rnbase_Domain_Model_DynamicTableInterface extends Sys25\RnBase\Domain\Model\DynamicTableInterface
+{
+}
+/** @deprecated */
+interface Tx_Rnbase_Domain_Model_RecordInterface extends Sys25\RnBase\Domain\Model\RecordInterface
+{
+}
+/** @deprecated */
+class Tx_Rnbase_Domain_Model_Data extends Sys25\RnBase\Domain\Model\DataModel implements Tx_Rnbase_Domain_Model_DataInterface
+{
+}
+/** @deprecated */
+class Tx_Rnbase_Domain_Model_Base extends Sys25\RnBase\Domain\Model\BaseModel implements Tx_Rnbase_Domain_Model_DomainInterface, Tx_Rnbase_Domain_Model_DynamicTableInterface, Tx_Rnbase_Domain_Model_RecordInterface
+{
+}
+/** @deprecated */
+trait Tx_Rnbase_Domain_Model_StorageTrait
+{
+    use Sys25\RnBase\Domain\Model\StorageTrait;
+}
+/** @deprecated */
+interface Tx_Rnbase_Domain_Repository_InterfaceSearch extends Sys25\RnBase\Domain\Repository\SearchInterface
+{
+}
+/** @deprecated */
+interface Tx_Rnbase_Domain_Repository_InterfacePersistence extends Sys25\RnBase\Domain\Repository\PersistenceInterface
+{
+}
+/** @deprecated */
+abstract class Tx_Rnbase_Domain_Repository_AbstractRepository extends Sys25\RnBase\Domain\Repository\AbstractRepository implements Tx_Rnbase_Domain_Repository_InterfaceSearch
+{
+}
+/** @deprecated */
+abstract class Tx_Rnbase_Domain_Repository_PersistenceRepository extends Sys25\RnBase\Domain\Repository\PersistenceRepository implements Tx_Rnbase_Domain_Repository_InterfacePersistence
+{
+}
+
+echo __LINE__.'<br>';
+/** @deprecated */
+abstract class Tx_Rnbase_Repository_AbstractRepository extends \Tx_Rnbase_Domain_Repository_AbstractRepository
+{
+}
+echo __LINE__.'<br>';