+++ cache/class.tx_rnbase_cache_Manager.php
@@ -22,8 +22,6 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  ***************************************************************/
 
-tx_rnbase::load('tx_rnbase_util_TYPO3');
-
 /**
  * This class provides access to caches.
  */