+++ ext_tables.php
@@ -1,10 +1,12 @@
 <?php
 
+use Sys25\RnBase\Utility\TYPO3;
+
 defined('TYPO3_MODE') or exit('Access denied.');
 
 if (TYPO3_MODE == 'BE') {
     // register rnbase dispatcher for modules before the extbase dispatcher
-    if (!tx_rnbase_util_TYPO3::isTYPO80OrHigher()) {
+    if (!TYPO3::isTYPO80OrHigher()) {
         tx_rnbase::load('Tx_Rnbase_Backend_ModuleRunner');
         $GLOBALS['TBE_MODULES']['_dispatcher'] = array_merge(
             ['Tx_Rnbase_Backend_ModuleRunner'],