+++ mod/base/class.tx_rnbase_mod_base_Lister.php
@@ -117,7 +117,7 @@ abstract class tx_rnbase_mod_base_Lister
     /**
      * Liefert den Service.
      *
-     * @return Tx_Rnbase_Domain_Repository_InterfaceSearch
+     * @return Sys25\RnBase\Domain\Repository\SearchInterface
      */
     abstract protected function getService();